# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Maquettiste editor API
  version: 0.2.0-phase2
  summary: The JSON API the Maquettiste editor's C# functions expose on the editor's own domain (SPEC Section 16), phase 2 subset.
  description: |
    The editor site's `_functions/` answer this API on the editor's domain (`http://maquettiste.localhost:8080` in local
    mode). Every handler is route parsing, the sign-in check and one call into `Maquettiste.Engine`
    (`docs/engineering/engine-design.md` section 15, `docs/engineering/phase2-design.md` section 3). The binding
    design is `docs/engineering/phase2-design.md`; this file is the contract the editor SPA, its generated client and its
    MSW mocks are built from.

    **Serialization.** Every body that mirrors an engine record is that record serialized by `System.Text.Json` with
    `JsonSerializerDefaults.Web` and no custom converters (host-contracts requirement 5): camelCase names, every
    property written, `null` included, enums as the kebab-case strings the engine declares, `TimeSpan` as
    `[-][d.]hh:mm:ss[.fffffff]`, `DateTimeOffset` as ISO 8601 with offset, and non-ASCII and HTML-sensitive characters
    escaped as `\uXXXX` (the default encoder). Computed public properties are serialized too (`ChangeSet.isEmpty`,
    `ValidationReport.hasErrors`). Schemas marked `x-engine-record` name the C# record they mirror; a change to that
    record is a change to this file in the same commit.

...
    and sends `json`, never `element`.

...

...

...

...

...
servers:
...
security:
  - sessionCookie: []
  - editorToken: []
...
paths:
...
  /api/model/index:
    get:
      tags: [model]
      operationId: getModelIndex
      summary: Summaries of every element
      description: |
        One summary per top-level element (sub-elements such as attributes and relation ends are not listed). The explorer,
        the command palette and every canvas start from this list.

        E5e: the response carries an `ETag`, a hash of every row's id, file hash and path and of the index format, so it
        changes exactly when a row does, across server restarts too. The index format constant is `maquettiste-index/e6`.

        `?locale=<tag>` (a declared locale; the default locale is the same as none) fills `displayName` from that locale's fallback
        chain (reference-types-seeds-localization.md section 3.8) from a per-locale display-name table the host builds once per
        snapshot; the ETag then also covers the locale, its chain and the hashes of the chain's shards. 400 for an undeclared locale. A request whose `If-None-Match` names it gets 304 with no
        body. `Cache-Control: no-cache` lets the browser keep the index and revalidate it on every use.
      x-maquettiste-role: viewer
      x-maquettiste-engine: ModelStore.GetIndexAsync
      x-engine-addition: E5e ModelReads.IndexTag(summaries)
      parameters:
        - $ref: '#/components/parameters/IfNoneMatch'
        - name: locale
          in: query
          description: A declared locale whose display names fill `displayName`.
          schema:
            type: string
            pattern: '^[A-Za-z]{2,8}(-[A-Za-z0-9]{1,8})*$'
      responses:
        '200':
          description: The index.
          headers:
            ETag:
              $ref: '#/components/headers/IndexETag'
            Cache-Control:
              $ref: '#/components/headers/NoCache'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ElementSummary'
              examples:
                billing:
                  summary: Part of the billing fixture's index
                  value:
                    - id: 01J92P0V01KDRN8GX5PGYCNKSX
                      kind: package
                      name: Billing
                      package: null
                      tags: []
                      category: null
                      stereotypes: []
                      hash: 28aadbb9e1c11cdd89539991fa4ced5d566312d0a62c7692bb7e5789cce923b2
                      path: .maquettiste/model/packages/billing.json
                    - id: 01J92P0V0FJ23CGSNKM7P1W5V7
                      kind: entity
                      name: Invoice
                      package: 01J92P0V01KDRN8GX5PGYCNKSX
                      tags:
                      - billing
                      category: 01J92P0V06TYE8P8990AV35A8K
                      stereotypes: 
                      - aggregate-root
                      - audited
                      - soft-delete
                      hash: cdac722e6a2f7418b8f02fb4ab7097f297bd46bc67f46d9f61c45dfda534749d
                      path: .maquettiste/model/entities/invoice.json
                    - id: 01J92P0V1BWHG0REWKSTR292RS
                      kind: relation
                      name: contains
                      package: 01J92P0V01KDRN8GX5PGYCNKSX
                      tags: []
                      category: null
                      stereotypes: []
                      hash: f9fdc699d65150052e25ef459180736b24a44f97c69cbefd09c9d12897fd36e9
                      path: .maquettiste/model/relations/contains.json
                      ends:
                      - entity: 01J92P0V0FJ23CGSNKM7P1W5V7
                        role: invoice
                      - entity: 01J92P0V0GWFR78HZH0P8Z3GY7
                        role: lines
                    - id: 01J92P0V1QRN2181XM2ZWE02W4
                      kind: database
                      name: main
                      package: null
                      tags: []
                      category: null
                      stereotypes: []
                      hash: 7fccf6c4f0985534e9089ca5b34664883add3de727699fb83d825ae1d0a638f5
                      path: .maquettiste/model/databases/main/database.json
                    - id: 01J92P0V2164SDBW687ZV6E1MV
                      kind: diagram
                      name: Billing overview
                      package: 01J92P0V01KDRN8GX5PGYCNKSX
                      tags: []
                      category: null
                      stereotypes: []
                      hash: a29e751129658af7503fb429afb3b382538a96bd8adee116e98b08a17152840c
                      path: .maquettiste/model/diagrams/billing-overview.json
                      memberCount: 8
        '304':
          description: The index still has the ETag given in `If-None-Match` (E5e).
          headers:
            ETag:
              $ref: '#/components/headers/IndexETag'
            Cache-Control:
              $ref: '#/components/headers/NoCache'
        '503':
          $ref: '#/components/responses/ModelUnavailable'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /api/model/elements/read:
    post:
      tags: [model]
      operationId: readElements
      summary: Up to 200 elements in one request
      description: |
        E5b: the documents of up to 200 element or sub-element ids, read from one model snapshot, for diagram and inspector
        loads (one request per chunk of ids instead of one per element). Each document comes back once, in the order its first
        id was asked for; ids that match nothing are listed in `missing`. More than 200 ids is `bad-request`.
      x-maquettiste-role: viewer
      x-engine-addition: E5b ModelStore.ReadElementsAsync(ids)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ElementReadRequest'
            example:
              ids: [01J92P0V0FJ23CGSNKM7P1W5V7, 01J92P0V1BWHG0REWKSTR292RS]
      responses:
        '200':
          description: The documents and the ids that matched nothing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ElementReadResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ModelUnavailable'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /api/model/elements:
    post:
      tags: [model]
      operationId: createElement
      summary: Create an element
      description: |
        The body is a model document of any kind. A missing top-level `id` is assigned by the engine; sub-element ids
        (attributes, relation ends, keys, columns) must be present, so the editor generates them (uppercase ULIDs). The
        engine writes the canonical file in the kind's conventional folder.
      x-maquettiste-role: editor
      x-maquettiste-engine: ModelStore.CreateAsync(json, ChangeSource.Editor)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewModelDocument'
            examples:
              entity:
                summary: A new entity with two attributes
                value:
                  kind: entity
                  name: Refund
                  package: 01J92P0V01KDRN8GX5PGYCNKSX
                  key:
                    attributes:
                    - 01M3P0000000000000000000A1
                  attributes:
                  - id: 01M3P0000000000000000000A1
                    name: id
                    type: uuid
                    required: true
                  - id: 01M3P0000000000000000000A2
                    name: amount
                    type: decimal
                    precision: 12
                    scale: 2
                    required: true
      responses:
        '201':
          description: Created. `current` is the new document.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Location:
              description: /api/model/elements/{id}
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveResult'
              examples:
                created:
                  summary: The entity was written
                  value:
                    outcome: saved
                    id: 01M3MNY0GJQHEAHZ3C98GZZ3Q3
                    hash: 90df3d3dc5d29b121aff75d34dfd3abc4706049ad84a583ee9d09abafd6e2bcc
                    current:
                      element:
                        kind: entity
                        package: 01J92P0V01KDRN8GX5PGYCNKSX
                        abstract: false
                        base: null
                        key:
                          attributes:
                          - 01M3P0000000000000000000A1
                          strategy: application
                        alternateKeys: []
                        attributes:
                        - type: uuid
                          required: true
                          default: null
                          defaultExpression: null
                          length: null
                          precision: null
                          scale: null
                          collection: false
                          unique: false
                          indexed: false
                          readOnly: false
                          immutable: false
                          derived: null
                          sensitive: null
                          validation: null
                          order: null
                          id: 01M3P0000000000000000000A1
                          name: id
                          displayName: null
                          pluralName: null
                          description: null
                          tags: []
                          category: null
                          stereotypes: []
                          properties: {}
                          generation: {}
                          source: null
                        - type: decimal
                          required: true
                          default: null
                          defaultExpression: null
                          length: null
                          precision: 12
                          scale: 2
                          collection: false
                          unique: false
                          indexed: false
                          readOnly: false
                          immutable: false
                          derived: null
                          sensitive: null
                          validation: null
                          order: null
                          id: 01M3P0000000000000000000A2
                          name: amount
                          displayName: null
                          pluralName: null
                          description: null
                          tags: []
                          category: null
                          stereotypes: []
                          properties: {}
                          generation: {}
                          source: null
                        $schema: ../../.schema/v1/entity.json
                        id: 01M3MNY0GJQHEAHZ3C98GZZ3Q3
                        name: Refund
                        displayName: null
                        pluralName: null
                        description: null
                        tags: []
                        category: null
                        stereotypes: []
                        properties: {}
                        generation: {}
                        source: null
                      path: .maquettiste/model/entities/refund.json
                      hash: 90df3d3dc5d29b121aff75d34dfd3abc4706049ad84a583ee9d09abafd6e2bcc
                      dependencyHash: 9abb16dafe7ecb690d835e0e3bc46870b3a42227e5be4dfefdaab13f8209459f
                      json:
                        $schema: ../../.schema/v1/entity.json
                        kind: entity
                        id: 01M3MNY0GJQHEAHZ3C98GZZ3Q3
                        name: Refund
                        package: 01J92P0V01KDRN8GX5PGYCNKSX
                        key:
                          attributes:
                          - 01M3P0000000000000000000A1
                        attributes:
                        - id: 01M3P0000000000000000000A1
                          name: id
                          type: uuid
                          required: true
                        - id: 01M3P0000000000000000000A2
                          name: amount
                          type: decimal
                          precision: 12
                          scale: 2
                          required: true
                      sidecarText: null
                    diagnostics: []
                    referrers: []
                    changes:
                      changed:
                      - id: 01M3MNY0GJQHEAHZ3C98GZZ3Q3
                        kind: entity
                        path: .maquettiste/model/entities/refund.json
                        hash: 90df3d3dc5d29b121aff75d34dfd3abc4706049ad84a583ee9d09abafd6e2bcc
                      deleted: []
                      source: editor
                      truncated: false
                      isEmpty: false
        '422':
          description: Invalid (JSON, schema or semantic validation); nothing was written.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveResult'
              examples:
                missingSubId:
                  summary: An attribute without an id
                  value:
                    outcome: invalid
                    id: 01M3MNY0GGD7TQPTYJFMGFFDCW
                    hash: null
                    current: null
                    diagnostics:
                    - rule: MQ1002
                      severity: error
                      message: '/attributes/0 required: Required properties ["id"] are not present'
                      elementId: 01M3MNY0GGD7TQPTYJFMGFFDCW
                      filePath: null
                      jsonPointer: /attributes/0
                      line: null
                      column: null
                    referrers: []
                    changes: null
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
...
  /api/model/batch:
    post:
      tags: [model]
      operationId: applyBatch
      summary: Atomic multi-element change
      description: |
        Refactorings, bulk edits, undo and redo, and pasted or proposed changes. The body is validated against
        `schemas/v1/batch.json` before any disk access, then the whole batch is validated in memory, each expected hash
        is checked, and every file is staged and renamed into place; on any failure nothing is written. Failed items carry
        their outcome; the other items report `saved` with no hash.
      x-maquettiste-role: editor
      x-maquettiste-engine: ModelStore.ParseBatch then ModelStore.ApplyBatchAsync(batch, ChangeSource.Editor)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '../../schemas/v1/batch.json'
            examples:
              createEntity:
                summary: Create one entity
                value:
                  operations:
                  - op: create
                    element:
                      kind: entity
                      name: Shipment
                      package: 01J92P0V01KDRN8GX5PGYCNKSX
      responses:
        '200':
          description: Every operation succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchResult'
        '409':
          description: An expected hash was stale (`conflict`) or a delete was refused (`referenced`); nothing was written.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchResult'
        '404':
          description: An update or delete named an unknown id; nothing was written.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchResult'
        '422':
          description: The batch did not parse (`BatchParseResult`, `batch` null) or an operation is invalid (`BatchResult`); nothing was written.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/BatchResult'
                  - $ref: '#/components/schemas/BatchParseResult'
              examples:
                invalidItem:
                  summary: A created entity without a key
                  value:
                    outcome: invalid
                    items:
                    - outcome: invalid
                      id: 01M3MNY0GZHRR68GZ2SZRGKF0D
                      hash: null
                      current: null
                      diagnostics:
                      - rule: MQ3005
                        severity: error
                        message: Entity 'Shipment' has no key; add 'key', or make it abstract or derived.
                        elementId: 01M3MNY0GZHRR68GZ2SZRGKF0D
                        filePath: .maquettiste/model/entities/shipment.json
                        jsonPointer: ''
                        line: 1
                        column: 1
                      referrers: []
                      changes: null
                    changes: null
                unparsable:
                  summary: An unknown op
                  value:
                    batch: null
                    diagnostics:
                    - rule: MQ1002
                      severity: error
                      message: '/operations/0/op enum: Value should match one of the values specified by the enum'
                      elementId: null
                      filePath: null
                      jsonPointer: /operations/0/op
                      line: 1
                      column: 22
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /api/model/references/{id}:
    parameters:
      - $ref: '#/components/parameters/ElementId'
    get:
      tags: [model]
      operationId: getReferences
      summary: Where used
      description: Every reference to the element or one of its sub-elements from another element.
      x-maquettiste-role: viewer
      x-maquettiste-engine: ModelStore.GetReferencesAsync (404 when ModelStore.GetElementAsync finds nothing)
      responses:
        '200':
          description: The references.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReferenceInfo'
              examples:
                invoice:
                  summary: Who references Invoice
                  value:
                    - fromElementId: 01J92P0V1T0J6RH4MY9H81NYB4
                      fromId: 01J92P0V1T0J6RH4MY9H81NYB4
                      jsonPointer: /entity
                      field: entity
                      toId: 01J92P0V0FJ23CGSNKM7P1W5V7
                    - fromElementId: 01J92P0V2164SDBW687ZV6E1MV
                      fromId: 01J92P0V2164SDBW687ZV6E1MV
                      jsonPointer: /members/1/element
                      field: element
                      toId: 01J92P0V0FJ23CGSNKM7P1W5V7
                    - fromElementId: 01J92P0V200XW93JQYSRCT2SE3
                      fromId: 01J92P0V200XW93JQYSRCT2SE3
                      jsonPointer: /entity
                      field: entity
                      toId: 01J92P0V0FJ23CGSNKM7P1W5V7
                    - fromElementId: 01J92P0V1BWHG0REWKSTR292RS
                      fromId: 01J92P0V1GMF7GJPA7981CH7YG
                      jsonPointer: /ends/0/entity
                      field: entity
                      toId: 01J92P0V0FJ23CGSNKM7P1W5V7
                    - fromElementId: 01J92P0V1ACKN3G6TK3NJDTM82
                      fromId: 01J92P0V1F439P4NH5F6HV5QM1
                      jsonPointer: /ends/1/entity
                      field: entity
                      toId: 01J92P0V0FJ23CGSNKM7P1W5V7
                    - fromElementId: 01J92P0V1DKHD5Q02DV6S5M53D
                      fromId: 01J92P0V1N727B9EM9FXPMKH7P
                      jsonPointer: /ends/1/entity
                      field: entity
                      toId: 01J92P0V0FJ23CGSNKM7P1W5V7
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthenticated'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: mq_session
      description: Set by `POST /api/session`. Host-only, HttpOnly, SameSite=Strict; protected with the host's data-protection keys; carries the user name, role, issue time and a fingerprint of the editor token.
    editorToken:
      type: http
      scheme: bearer
      description: '`Authorization: Bearer <MAQUETTISTE_EDITOR_TOKEN>`, the secret site variable. For the CLI, scripts and tests. Compared in fixed time. In local mode a loopback or `MAQUETTISTE_LOCAL_PEERS` caller with a local `Host` and no forwarding header needs neither scheme and is the local admin (see the authentication notes in `info`).'
  parameters:
...
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      schema:
        type: string
  headers:
...
    IndexETag:
      description: The index's hash as a strong ETag (E5e).
      schema:
        type: string
        pattern: '^"[0-9a-f]{64}"$'
    NoCache:
      description: '`no-cache`: the client may keep the response but revalidates it with `If-None-Match` before each use (E5e).'
      schema:
        type: string
        const: no-cache
...